The app can now remember hostnames and ports that the user has typed in.
This is done by keeping a record of all visited hosts and their port numbers
(hosts are only saved once a successful connection has been made) and if an
existing host exists but is connected to on a new port number, the portnumber
of the existing host will be changes. You do not get two identical hostnames
with different port numbers.

Port numbers are mostly updated when a hostname is typed in, is selected,
appears as if by magic in the host field.

When the application starts the fields are initialised to the host that was
connected to last.

The change in the nib is a delegate connection from the host combo box to the
BuildGrowlerController.