An intermediary commit. This won't run due to bits that are missing.
The commit implementes the updating of the username and password fields when
the host name changes. I currently store a bool indicating whether a password
is associated with a host, the password (will be) stored in the keychain. I've
decided that there is no reason to store that boolean however, but instead
assume that if there is a username, then there is also a password, even if it
is blank. I think that makes better sense than the boolean.