This adds a support class for building frameworks to setup.py
The framework BTUtils, which contains a selection of files from the Adium
utilities package. I might add more stuff in the future. I opted not to go for
the http://sourceforge.net/projects/keychain/ framework as it just seemed a bit
overkill. Of course I was not able to wield my scalpel as expertly as I had
wished and so quite a few bits of other support files are required in addition
to AIKeychain. AIKeychain does have pretty good documentation, which is a bonus.
Although so does the keychain framework on Sourceforge.

Added config items to setup.py to build the BGUtils framework.