Added some new logic for password changes.
A new class PasswordController (which the password field is an instance of) (and
this class is a subclass of NSSecureTextField) has been added which deals with
the password change detection. It also clears the passwordfield whenever the
user enters it (ie it receives focus) and gets set to eight dots when the user
exits (when editing finishes). The password is only ever stored in the field
while the user has focus on it, otherwise it is put aside and stored in an
NSWiredString.